+++ Common/src/main/java/net/merchantpug/bovinesandbuttercups/platform/services/IPlatformHelper.java
@@ -38,6 +38,4 @@ public interface IPlatformHelper {
 
     Map<Block, Block> getPottedBlockMap();
 
-    void addMoobloomFlowerBlockProperties(StateDefinition.Builder<Block, BlockState> builder);
-
 }