+++ Fabric/src/main/java/net/merchantpug/bovinesandbuttercups/platform/FabricPlatformHelper.java
@@ -81,9 +81,4 @@ public class FabricPlatformHelper implements IPlatformHelper {
         return FlowerPotBlock.POTTED_BY_CONTENT;
     }
 
-    @Override
-    public void addMoobloomFlowerBlockProperties(StateDefinition.Builder<Block, BlockState> builder) {
-        if (isModLoaded("stackableflowers"))
-            builder.add(BlockStateProperties.FLOWER_AMOUNT);
-    }
 }