+++ CHANGELOG.md
@@ -1,2 +1,2 @@
-## Integration
-- [FABRIC/QUILT] Fixed a crash with Stackable Flowers, as Bovines has different blockstates. #20
+## Bugfixes
+- [FABRIC/QUILT] Fixed a crash with other mods that modify FlowerBlock's blockstates. #20